requirements.txt
pytest>=7.4.0
pytest-html>=3.2.0
bandit>=1.7.5
semgrep>=1.45.0
pylint>=3.0.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
qrcode[pil]>=7.4.2
